SRB2: USHORT replaced
git-svn-id: svn://code.srb2.org/SRB2/trunk@4625 d18046d3-e9af-4449-a702-ac2b6c676ba2

+++ src/r_draw.c
@@ -545,7 +545,7 @@ void R_FillBackScreen(void)
 
 
 */
-void R_VideoErase(unsigned ofs, INT32 count)
+void R_VideoErase(size_t ofs, INT32 count)
 {
 	// LFB copy.
 	// This might not be a good idea if memcpy
